todo_app/todo_api/models.py
from django.db import models
from django.contrib.auth.models import User
from hashid_field import HashidAutoField

class Todo(models.Model):
    id = HashidAutoField(primary_key=True)
    title = models.CharField(max_length=255)
    description = models.TextField(blank=True)
    is_completed = models.BooleanField(default=False)
    created_at = models.DateTimeField(auto_now_add=True)
    updated_at = models.DateTimeField(auto_now=True)
    user = models.ForeignKey(User, on_delete=models.CASCADE)
    
    def __str__(self):
        return self.title