todo_app/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.13
django-allauth==0.62.1
django-cors-headers==4.3.1
django-filter==24.2
django-hashid-field==3.4.1
django-postgres-extra==2.0.8
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-model-serializer==0.0.2
drf-rw-serializers==1.1.1
hashids==1.3.1
idna==3.7
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
urllib3==2.2.1
